requirements.txt
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.25.0
scikit-learn>=1.3.0
tqdm>=4.66.0
# PDF: pypdf — компактный пакет; при таймаутах PyPI сначала ставьте его.
pypdf>=4.0.0
# Опционально (быстрее на больших PDF): pip install pymupdf
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.1
python-docx>=1.1.0
gradio>=5.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
requests>=2.32.0
pymupdf>=1.24.0
